Move unittests into separate directory
